#703 #704 Refactor and editor fix
