+++ WMIAdventure/frontend/src/js/pages/MainMenu/MainMenu.js
@@ -6,7 +6,7 @@ import adventureImg from '../../../assets/images/adventure-image.png';
 import battleImg from '../../../assets/images/battle-image.png';
 import rankingImg from '../../../assets/images/ranking-image.png';
 import StyledWrapper from './StyledWrapper';
-import Navbar from "../../components/global/atoms/Navbar";
+import Navbar from "../../components/global/molecules/Navbar";
 
 
 function MainMenu() {
