+++ WMIAdventure/frontend/src/js/pages/UserRegistration/UserRegistrationPage.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {Helmet} from "react-helmet";
 import StyledWrapper from "./StyledWrapper";
 import UserRegistration from "../../components/auth/organisms/UserRegistration/UserRegistration";
-import Navbar from "../../components/global/atoms/Navbar";
+import Navbar from "../../components/global/molecules/Navbar";
 
 function UserRegistrationPage() {
     return (
